+++ src/components/react-hook-form/SelectItem/SelectItemFormFieldInput.tsx
@@ -10,7 +10,10 @@ import { useId } from '../../util'
 import { useFieldError } from '../util/useFieldError'
 import { replacePlaceholderColorWithTextColor } from './replacePlaceholderColorWithTextColor'
 
-export type SelectItemInputProps<Item, SelectedItem extends Item | null> = {
+export type SelectItemFormFieldInputProps<
+  Item,
+  SelectedItem extends Item | null,
+> = {
   name: string
   label: string
   placeholder: string
@@ -26,7 +29,10 @@ export type SelectItemInputProps<Item, SelectedItem extends Item | null> = {
 } & RequiredProps &
   HideRequiredProps
 
-export function SelectItemInput<Item, SelectedItem extends Item | null>({
+export function SelectItemFormFieldInput<
+  Item,
+  SelectedItem extends Item | null,
+>({
   name,
   label,
   placeholder,
@@ -41,7 +47,7 @@ export function SelectItemInput<Item, SelectedItem extends Item | null>({
   className,
   required,
   hideRequired,
-}: SelectItemInputProps<Item, SelectedItem>) {
+}: SelectItemFormFieldInputProps<Item, SelectedItem>) {
   const id = useId()
   const componentRef = useRef<HTMLDivElement | null>(null)
