+++ src/components/react-hook-form/SelectItem/__test__/replacePlaceholderColorWithTextColor.test.ts
