+++ src/components/react-hook-form/SelectItem/replacePlaceholderColorWithTextColor.ts
