--- src/components/react-hook-form/async/index.ts
@@ -1,4 +1,0 @@
-export * from './replacePlaceholderColorWithTextColor'
-export * from './SelectItemFormField'
-export * from './SelectItemFormFieldDialog'
-export * from './SelectItemInput'
