+++ src/components/react-hook-form/index.ts
@@ -15,5 +15,6 @@ export * from './ToggleSwitchFormField'
 export * from './RadioFormField'
 export * from './SearchFormField'
 export * from './useHandleSubmit'
-export * from './async'
+export * from './SelectItemFormField'
+export * from './SelectItem'
 export * from './util'
